requirements.txt
appdirs==1.4.3
certifi==2019.11.28
distlib==0.3.0
filelock==3.0.12
importlib-metadata==1.5.0
numpy==1.18.1
Pillow==7.0.0
pyexiv2==2.1.0
rawkit==0.6.0
six==1.14.0
zipp==3.1.0
